% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-cron_profiles_fcts.R
\name{cies_profile_build}
\alias{cies_profile_build}
\title{Build simplified company profile table}
\usage{
cies_profile_build(con)
}
\arguments{
\item{con}{Connexion DBI}
}
\description{
Prend la table `cies_profile_orig` et crée/écrase `cies_profile_build`
avec seulement les colonnes utiles.
}
