% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-cron_profiles_fcts.R
\name{fmp_company_screener}
\alias{fmp_company_screener}
\title{Get symbols from FMP company screener}
\usage{
fmp_company_screener(
  key_fmp_api,
  exchange = "NASDAQ",
  country = "US",
  limit = 10000
)
}
\arguments{
\item{key_fmp_api}{Clé API Financial Modeling Prep}

\item{exchange}{Exchange (ex: "NASDAQ", "NYSE", "AMEX")}

\item{country}{Pays (ex: "US")}

\item{limit}{Nombre max de résultats (max recommandé: 10000)}
}
\value{
Un data.frame ou NULL
}
\description{
Récupère une liste de tickers filtrés (exchange, market cap, etc.)
depuis l’API Financial Modeling Prep.
}
