% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-cron_split_fcts.R
\name{get_pending_splits}
\alias{get_pending_splits}
\title{Récupérer les splits "pending" à traiter}
\usage{
get_pending_splits(con, horizon_days = 2)
}
\arguments{
\item{con}{Objet `DBIConnection`}

\item{horizon_days}{Nombre de jours à regarder dans le futur (par défaut 2)}
}
\value{
`data.frame` contenant les splits pending triés par `split_date`
}
\description{
Retourne les splits dont le statut est `"pending"` et dont la date de split
est inférieure ou égale à `CURRENT_DATE + horizon_days`.
}
