% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-api_fcts.R
\name{init_log}
\alias{init_log}
\title{Initialiser l'accumulateur de log}
\usage{
init_log()
}
\value{
Invisiblement NULL.
}
\description{
Vide l'accumulateur `log_env$log_data` (data.frame vide). Appelé en début
de `0-plan.R` avant l'exécution des étapes.
}
