% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance4.R
\name{mod_finance4_server}
\alias{mod_finance4_server}
\title{Serveur du module analyse boursière v4}
\usage{
mod_finance4_server(id, con)
}
\arguments{
\item{id}{Identifiant du module}

\item{con}{Connexion DBI active (PostgreSQL)}
}
\description{
Serveur du module analyse boursière v4
}
