% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance5.R
\name{mod_finance5_server}
\alias{mod_finance5_server}
\title{Serveur du module analyse boursière v5}
\usage{
mod_finance5_server(id, con)
}
\arguments{
\item{id}{Identifiant du module}

\item{con}{Connexion DBI active (PostgreSQL)}
}
\description{
Serveur du module analyse boursière v5
}
