% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance6.R
\name{mod_finance6_server}
\alias{mod_finance6_server}
\title{Serveur du module analyse boursière v6}
\usage{
mod_finance6_server(id, con)
}
\arguments{
\item{id}{Identifiant du module}

\item{con}{Connexion DBI active (PostgreSQL)}
}
\description{
Serveur du module analyse boursière v6
}
