% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance6.R
\name{mod_finance6_ui}
\alias{mod_finance6_ui}
\title{UI complète du module analyse boursière v6}
\usage{
mod_finance6_ui(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::navset_bar`
}
\description{
Inclut le sélecteur de compagnie partagé dans la navbar (visible dans tous les onglets).
}
