% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance6.R
\name{mod_finance6_ui1}
\alias{mod_finance6_ui1}
\title{Onglet Filtres du module analyse boursière v6}
\usage{
mod_finance6_ui1(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::nav_panel`
}
\description{
Onglet Filtres du module analyse boursière v6
}
