% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance6.R
\name{mod_finance6_ui2}
\alias{mod_finance6_ui2}
\title{Onglet check quality du module analyse boursière v6}
\usage{
mod_finance6_ui2(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::nav_panel`
}
\description{
Section QUALITÉ uniquement : Marges | ROA/ROE + Croissance | Structure financière.
}
