% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance6.R
\name{mod_finance6_ui3}
\alias{mod_finance6_ui3}
\title{Onglet check price du module analyse boursière v6}
\usage{
mod_finance6_ui3(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::nav_panel`
}
\description{
Section PRIX & INTÉRÊT : graphique buy/sell (2/3 gauche, pleine hauteur) +
Valorisation et Dividendes empilés (1/3 droite).
}
