% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance.R
\name{mod_finance_server}
\alias{mod_finance_server}
\title{Serveur du module analyse boursière}
\usage{
mod_finance_server(id, con)
}
\arguments{
\item{id}{Identifiant du module (doit correspondre à celui passé à `mod_finance_ui`)}

\item{con}{Connexion DBI active (PostgreSQL)}
}
\description{
Serveur du module analyse boursière
}
