% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance.R
\name{mod_finance_ui}
\alias{mod_finance_ui}
\title{UI complète du module analyse boursière (raccourci)}
\usage{
mod_finance_ui(id)
}
\arguments{
\item{id}{Identifiant du module (chaîne de caractères)}
}
\value{
Un `bslib::navset_bar`
}
\description{
Combine `mod_finance_ui1()` et `mod_finance_ui2()` dans un `bslib::navset_bar()`
prêt à être inséré dans un conteneur de page (ex: `bslib::page_fluid()`).
}
