% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance.R
\name{mod_finance_ui1}
\alias{mod_finance_ui1}
\title{Onglet Filtres du module analyse boursière}
\usage{
mod_finance_ui1(id)
}
\arguments{
\item{id}{Identifiant du module (chaîne de caractères)}
}
\value{
Un `bslib::nav_panel`
}
\description{
Retourne un `bslib::nav_panel` "Filtres" à insérer dans un `navset_*` parent.
Utiliser le même `id` que `mod_finance_ui2()` et `mod_finance_server()`.
}
