% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_logs.R
\name{mod_logs_server}
\alias{mod_logs_server}
\title{Server du module de consultation des logs}
\usage{
mod_logs_server(id, con)
}
\arguments{
\item{id}{Identifiant du module (Shiny).}

\item{con}{Connexion DBI PostgreSQL.}
}
\value{
Invisiblement NULL.
}
\description{
Server du module de consultation des logs
}
