% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_logs.R
\name{mod_logs_ui}
\alias{mod_logs_ui}
\title{UI du module de consultation des logs}
\usage{
mod_logs_ui(id)
}
\arguments{
\item{id}{Identifiant du module (Shiny).}
}
\value{
Une UI Shiny (card bslib).
}
\description{
UI du module de consultation des logs
}
