% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market.R
\name{mod_market_ui}
\alias{mod_market_ui}
\title{UI du module de vue marché}
\usage{
mod_market_ui(id)
}
\arguments{
\item{id}{Identifiant du module (Shiny).}
}
\value{
Une UI Shiny (card bslib).
}
\description{
UI du module de vue marché
}
