% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-cron_earning_cal_fcts.R
\name{update_cies_order}
\alias{update_cies_order}
\title{Mettre à jour le fichier de suivi des publications}
\usage{
update_cies_order(earning_calendar, save_path)
}
\arguments{
\item{earning_calendar}{data.frame contenant au minimum les colonnes `symbol` et `date`}

\item{save_path}{Nom de l'objet S3 (ex: `"data/cies_order.rds"`)}
}
\value{
Invisiblement, le data.frame `cies_order` mis à jour.
}
\description{
Ajoute au fichier `cies_order.rds` (sur S3) les nouvelles compagnies présentes
dans le calendrier de publications mais absentes du suivi actuel.
Les compagnies déjà présentes ne sont pas modifiées.
}
