% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-cron_split_fcts.R
\name{fmp_splits_calendar_get}
\alias{fmp_splits_calendar_get}
\title{Get stock splits calendar from Financial Modeling Prep API (with date range)}
\usage{
fmp_splits_calendar_get(fmp_api_key, from = NULL, to = NULL)
}
\arguments{
\item{fmp_api_key}{Clé API Financial Modeling Prep}

\item{from}{Date de début au format "YYYY-MM-DD"}

\item{to}{Date de fin au format "YYYY-MM-DD"}
}
\value{
Un data.frame issu du JSON retourné par l’API, ou `NULL` si l’appel échoue.
}
\description{
Récupère le calendrier des splits boursiers pour tous les tickers depuis l’API Financial Modeling Prep,
en filtrant sur une plage de dates.
}
