% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance4.R
\name{mod_finance4_ui}
\alias{mod_finance4_ui}
\title{UI complète du module analyse boursière v4}
\usage{
mod_finance4_ui(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::navset_bar`
}
\description{
UI complète du module analyse boursière v4
}
