% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance5.R
\name{mod_finance5_ui}
\alias{mod_finance5_ui}
\title{UI complète du module analyse boursière v5}
\usage{
mod_finance5_ui(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::navset_bar`
}
\description{
UI complète du module analyse boursière v5
}
