% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance5.R
\name{mod_finance5_ui1}
\alias{mod_finance5_ui1}
\title{Onglet Filtres du module analyse boursière v5}
\usage{
mod_finance5_ui1(id)
}
\arguments{
\item{id}{Identifiant du module}
}
\value{
`bslib::nav_panel`
}
\description{
Onglet Filtres du module analyse boursière v5
}
