% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_finance.R
\name{mod_finance_ui}
\alias{mod_finance_ui}
\title{UI du module analyse boursière}
\usage{
mod_finance_ui(id)
}
\arguments{
\item{id}{Identifiant du module (chaîne de caractères)}
}
\value{
`tagList` contenant deux `bslib::nav_panel`
}
\description{
Retourne un `tagList` de deux `nav_panel` (Filtres + Résultats) à insérer
dans un `bslib::page_navbar()`.
}
